+++ Engine/Source/Tempest/caabox.cpp
@@ -5,7 +5,7 @@
 
 namespace NTempest {
 
-  CAaBox CAaBox::Bounding(const C3Vector *vectors, unsigned long count) {
+  CAaBox CAaBox::Bounding(const C3Vector *vectors, DWORD count) {
     ASSERT(vectors != 0);
 
     CAaBox extents;
@@ -13,7 +13,7 @@ namespace NTempest {
       extents.b = vectors[0];
       extents.t = vectors[0];
 
-      for (unsigned long i = 1; i < count; ++i) {
+      for (DWORD i = 1; i < count; ++i) {
         extents.b = C3Vector(
             vectors[i].x <= extents.b.x ? vectors[i].x : extents.b.x, vectors[i].y <= extents.b.y ? vectors[i].y : extents.b.y,
             vectors[i].z <= extents.b.z ? vectors[i].z : extents.b.z
@@ -25,7 +25,7 @@ namespace NTempest {
     return extents;
   }
 
-  CAaBox CAaBox::Bounding(const CDynTable<unsigned long> &index, const CDynTable<C3Vector> &vects) {
+  CAaBox CAaBox::Bounding(const CDynTable<DWORD> &index, const CDynTable<C3Vector> &vects) {
     ASSERT(index.IsValid() && vects.IsValid());
 
     CAaBox extents;
@@ -33,7 +33,7 @@ namespace NTempest {
       extents.b = vects[index[0]];
       extents.t = vects[index[0]];
 
-      for (unsigned long i = 1; i < index.Used(); ++i) {
+      for (DWORD i = 1; i < index.Used(); ++i) {
         extents.b = C3Vector::Min(extents.b, vects[index[i]]);
         extents.t.Maximize(vects[index[i]]);
       }
